+++ lib/islamic_text.dart
@@ -1,5 +1,2 @@
-/// A Calculator.
-class Calculator {
-  /// Returns [value] plus 1.
-  int addOne(int value) => value + 1;
-}
+export 'src/widgets/islamic_text_widget.dart';
+export 'src/ligature/ligature_replacer.dart';